requirements.txt
azure-storage-blob
pandas
pyarrow
requests
